+++ src/main/java/com/ducky/command/AddTaskCommand.java
@@ -3,9 +3,9 @@ package com.ducky.command;
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 
-import com.ducky.logic.DuckyException;
-import com.ducky.logic.Storage;
-import com.ducky.logic.TaskList;
+import com.ducky.common.DuckyException;
+import com.ducky.common.Storage;
+import com.ducky.common.TaskList;
 import com.ducky.task.DeadlineTask;
 import com.ducky.task.EventTask;
 import com.ducky.task.TaskType;
